Simplify, document, rename the RATS frame to FastADT
