+++ src/instamatic/experiments/__init__.py
@@ -3,6 +3,6 @@ from __future__ import annotations
 from .autocred import experiment as autocRED
 from .cred import experiment as cRED
 from .cred_tvips import experiment as cRED_tvips
-from .rats import experiment as rats
+from .fast_adt import experiment as fast_adt
 from .red import experiment as RED
 from .serialed import experiment as serialED
